Move ManagedTsService back
