+++ timelock-agent/src/main/java/com/palantir/timelock/paxos/DbBoundTimestampCreator.java
@@ -24,7 +24,7 @@ import com.palantir.atlasdb.AtlasDbConstants;
 import com.palantir.atlasdb.factory.ServiceDiscoveringAtlasSupplier;
 import com.palantir.atlasdb.spi.KeyValueServiceConfig;
 import com.palantir.atlasdb.timelock.paxos.DelegatingManagedTimestampService;
-import com.palantir.timestamp.ManagedTimestampService;
+import com.palantir.atlasdb.timelock.paxos.ManagedTimestampService;
 import com.palantir.timestamp.TimestampManagementService;
 import com.palantir.timestamp.TimestampService;
