+++ timelock-impl/src/main/java/com/palantir/atlasdb/timelock/paxos/DelegatingManagedTimestampService.java
@@ -18,7 +18,6 @@ package com.palantir.atlasdb.timelock.paxos;
 import javax.ws.rs.QueryParam;
 
 import com.google.common.base.Preconditions;
-import com.palantir.timestamp.ManagedTimestampService;
 import com.palantir.timestamp.TimestampManagementService;
 import com.palantir.timestamp.TimestampRange;
 import com.palantir.timestamp.TimestampService;
