+++ timelock-impl/src/main/java/com/palantir/atlasdb/timelock/paxos/ManagedTimestampService.java
@@ -13,7 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.palantir.timestamp;
+package com.palantir.atlasdb.timelock.paxos;
+
+import com.palantir.timestamp.TimestampManagementService;
+import com.palantir.timestamp.TimestampService;
 
 public interface ManagedTimestampService extends TimestampService, TimestampManagementService {
 }
