+++ timelock-impl/src/test/java/com/palantir/atlasdb/timelock/paxos/DelegatingManagedTimestampServiceTest.java
@@ -22,7 +22,6 @@ import static org.mockito.Mockito.verify;
 
 import org.junit.Test;
 
-import com.palantir.timestamp.ManagedTimestampService;
 import com.palantir.timestamp.TimestampManagementService;
 import com.palantir.timestamp.TimestampService;
